requirements.txt
fastapi
uvicorn

sqlalchemy

"passlib[bcrypt]"
python-multipart
"python-jose[cryptography]"
python-dotenv
python-decouple
alembic

aiofiles
jinja2
